final commit

+++ BlogApplicatie/controllers/BlogController.php
@@ -4,6 +4,7 @@ namespace app\controllers;
 
 use app\components\WebUser;
 use app\models\Comment;
+use Codeception\PHPUnit\Constraint\Page;
 use Yii;
 use app\models\Blog;
 use app\models\BlogSearch;
@@ -36,6 +37,7 @@ class BlogController extends Controller
      * Lists all Blog models.
      * @return mixed
      */
+    //action to run the blog index page
     public function actionIndex()
     {
         $searchModel = new BlogSearch();
@@ -55,6 +57,7 @@ class BlogController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
+    //action to run the page to view a blog individual
     public function actionView($id)
     {
         $model = Blog::findOne($id);
@@ -75,33 +78,45 @@ class BlogController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
+    //action to run the page to create a blog
     public function actionCreate()
     {
 
         $model = new Blog();
 
         $user = WebUser::findOne(Yii::$app->getUser()->id);
-        $a_id = $user->getId();
+
         $date = date('Y-m-d H:i:s');
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
 
-            $model->file = UploadedFile::getInstance($model, 'file');
-            $imageName = $model->file->getBaseName() . random_int(1, 100);
-            $model->attachment = 'uploads/' . $imageName . '.' . $model->file->extension;
-            $save = $model->attachment;
+        if(!$user) {
+            return $this->redirect('error');
+        } elseif($user->getAccessLevel() >= 16) {
 
-            $model->file->saveAs($save);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                if(UploadedFile::getInstance($model, 'file' !== NULL)) {
+                    $model->file = UploadedFile::getInstance($model, 'file');
+                    $imageName = $model->file->getBaseName() . random_int(1, 100);
+                    $model->attachment = 'uploads/' . $imageName . '.' . $model->file->extension;
+                    $save = $model->attachment;
 
-            $model->save();
+                    $model->file->saveAs($save);
+                }
 
-            return $this->redirect(['view', 'id' => $model->id]);
+                $model->save();
+
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+            return $this->render('create', [
+                'model' => $model,
+                'user' => $user,
+                'date' => $date
+            ]);
+        } else {
+            return $this->render('error');
         }
 
-        return $this->render('create', [
-            'model' => $model,
-            'user' => $user,
-            'date' => $date,
-        ]);
+
+
     }
 
     /**
@@ -111,22 +126,70 @@ class BlogController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
+    //action to run the page to update a blog
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+
         $user = WebUser::findOne(Yii::$app->getUser()->id);
-        $a_id = $user->getId();
         $date = date('Y-m-d H:i:s');
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+
+        if (!$user) {
+            return $this->redirect('error');
+        } elseif ($user->getAccessLevel() >= 98) {
+
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                if (UploadedFile::getInstance($model, 'file' !== NULL)) {
+                    $model->file = UploadedFile::getInstance($model, 'file');
+                    $imageName = $model->file->getBaseName() . random_int(1, 100);
+                    $model->attachment = 'uploads/' . $imageName . '.' . $model->file->extension;
+                    $save = $model->attachment;
+
+                    $model->file->saveAs($save);
+                }
+
+                $model->save();
+
+                return $this->redirect(['view', 'id' => $model->id]);
+
+            }
+            return $this->render('update', [
+                'model' => $model,
+                'user' => $user,
+                'date' => $date
+            ]);
+        }   elseif ($user->getAccessLevel() >= 16) {
+                if ($model->author_id == $user->id) {
+
+                    if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                        if (UploadedFile::getInstance($model, 'file' !== NULL)) {
+                            $model->file = UploadedFile::getInstance($model, 'file');
+                            $imageName = $model->file->getBaseName() . random_int(1, 100);
+                            $model->attachment = 'uploads/' . $imageName . '.' . $model->file->extension;
+                            $save = $model->attachment;
+
+                            $model->file->saveAs($save);
+                        }
+
+                        $model->save();
+
+                        return $this->redirect(['view', 'id' => $model->id]);
+                    }
+                    return $this->render('update', [
+                        'model' => $model,
+                        'user' => $user,
+                        'date' => $date
+                    ]);
+                } else {
+                    return $this->render("error");
+                }
+            } else {
+                return $this->render('error');
+            }
+
         }
 
-        return $this->render('update', [
-            'model' => $model,
-            'user' => $user,
-            'date' => $date,
-        ]);
-    }
+
 
     /**
      * Deletes an existing Blog model.
@@ -135,12 +198,43 @@ class BlogController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
+    //action to run the page to delete a blog
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        $searchModel = new BlogSearch();
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        $models = Blog::find()->all();
 
-        return $this->redirect(['index']);
+
+        $user = WebUser::findOne(Yii::$app->getUser()->id);
+        if(!$user) {
+            return $this->redirect('error');
+        }elseif($user->accessLevel() >= 98) {
+            $this->findModel($id)->delete();
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+                'models' => $models
+            ]);
+        } elseif($user->getAccessLevel() >= 16) {
+            if($this->findModel($id)->author_id == $user->id) {
+                $this->findModel($id)->delete();
+                return $this->render('index', [
+                    'searchModel' => $searchModel,
+                    'dataProvider' => $dataProvider,
+                    'models' => $models
+                ]);
+            } else {
+                return $this->render("error");
+            }
+
+        } else {
+            return $this->render('error');
+        }
+
+
     }
+
 
     /**
      * Finds the Blog model based on its primary key value.
@@ -149,6 +243,7 @@ class BlogController extends Controller
      * @return Blog the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
+    //method to find a model
     protected function findModel($id)
     {
         if (($model = Blog::findOne($id)) !== null) {
@@ -168,7 +263,7 @@ class BlogController extends Controller
         }
 
     }
-
+    //action to delete a comment
     public function actionDeleteComment($id) {
         $comment = Comment::findOne($id);
         $comment->delete();
